// generated/src/app/api-models/custserv-last-activity.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CustservLastActivity {
  claimsUpdateDate: Date;
  claimsAdjUpdateDate: Date;
  authUpdateDate: Date;
  demographicsUpdateDate: Date;
  pcpUpdateDate: Date;
  terminateUpdateDate: Date;
  reinstateUpdateDate: Date;
  idCardUpdateDate: Date;
  letterUpdateDate: Date;
  seqMembId: number;
}